backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend server for the new Bot-in-a-Box V2."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.121.0",
    "uvicorn[standard]==0.38.0",
    "python-multipart==0.0.20",
    "azure-storage-blob==12.27.0",
    "azure-search-documents==11.6.0",
    "azure-identity==1.25.1",
    "openai==1.109.1",
    "python-docx==1.2.0",
    "pandas==2.2.3",
    "openpyxl==3.1.5",
    "requests==2.32.5",
    "aiofiles==25.1.0",
    "PyJWT==2.10.1",
    "cryptography==46.0.3",
    "llama-index>=0.14.7",
    "llama-index-core>=0.14.7",
    "llama-index-embeddings-azure-openai==0.4.1",
    "llama-index-llms-azure-openai==0.4.2",
    "llama-index-vector-stores-azureaisearch>=0.4.2",
    "filelock>=3.20.0",
    "PyMuPDF==1.26.5",
    "reportlab>=4.0.0",
    "starlette==0.49.1",
    "pip==25.3",
    "pytest>=9.0.1",
    "markitdown[docx]==0.1.3",
    "opentelemetry-api>=1.10.0",
    "opentelemetry-sdk>=1.10.0",
    "azure-monitor-opentelemetry",
    "azure-monitor-opentelemetry-exporter",
    "opencensus>=0.11.2",
    "opencensus-ext-azure>=1.0.0",
    "python-dotenv>=1.0.0",
    "azure-identity-broker>=1.3.0",
    "tiktoken>=0.7.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["server*", "Service*"]
